backup media to telegram grp/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
logging==0.4.9.6
